Clean exit (exit status==0) when vde commands receive SIGTERM

+++ trunk/vde-2/src/vde_switch/vde_switch.c
@@ -473,7 +473,10 @@ static void sig_handler(int sig)
 	printlog(LOG_ERR,"Caught signal %d, cleaning up and exiting", sig);
 	cleanup();
 	signal(sig, SIG_DFL);
-	kill(getpid(), sig);
+	if (sig == SIGTERM)
+		_exit(0);
+	else
+		kill(getpid(), sig);
 }
 
 static void setsighandlers()
