+++ trunk/vde-2/src/vdeq.c
@@ -132,7 +132,10 @@ static void sig_handler(int sig)
 	fprintf(stderr,"%s: Caught signal %d, cleaning up and exiting\n", vdeqname, sig);
 	cleanup();
 	signal(sig, SIG_DFL);
-	kill(getpid(), sig);
+	if (sig == SIGTERM)
+		_exit(0);
+	else
+		kill(getpid(), sig);
 }
 
 static void setsighandlers()
